--- src/pages/settings/Payments/PaymentsPage/index.native.js
@@ -1,3 +1,0 @@
-import BasePaymentsPage from './BasePaymentsPage';
-
-export default BasePaymentsPage;
