+++ src/pages/settings/Wallet/WalletPage/index.native.js
@@ -1,0 +1,3 @@
+import BaseWalletPage from './BaseWalletPage';
+
+export default BaseWalletPage;
