+++ src/pages/settings/Wallet/WalletPage/paymentsPagePropTypes.js
